requirements.txt
Django==5.1.4
Markdown==3.4.1
PyJWT==1.7.1
asgiref==3.8.1
black==23.1.0
click==8.1.3
darker==1.7.0
django-cors-headers==4.6.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pytz==2024.2
setuptools==75.1.0
sqlparse==0.5.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
